Default to i32 for channel and difftime
This is better for the default, because Swabian channel is
(theoretically) 32-bit and Swabian-derived difftime could be greater
than 16-bit limit. BH and PQ values are 8 or 16 bit and fit in i32 with
no issue.

There is no glaring reason to suspect that using 32-bit will be worse
for performance, so let's use the convenient types until and unless
something is proved by benchmarking. (The reason for the previous 16-bit
default was mostly that only BH and PQ data had been considered.)

+++ include/libtcspc/common.hpp
@@ -48,12 +48,12 @@ struct default_data_traits {
     /**
      * \brief Channel number type.
      */
-    using channel_type = std::int16_t; // TODO int32_t
+    using channel_type = std::int32_t;
 
     /**
      * \brief Difference time type.
      */
-    using difftime_type = std::uint16_t; // TODO int32_t
+    using difftime_type = std::int32_t;
 
     /**
      * \brief Type of datapoint for histogramming.
